подключил компонент списка жанров напрямую к Store
