+++ src/components/genre-list/genre-list.jsx
@@ -1,7 +1,9 @@
 import React from 'react';
 import PropTypes from 'prop-types';
+import {connect} from 'react-redux';
+import {ActionCreator} from '../../store/action';
 
-const GenreList = ({genres, selectedGenreTab, handleGenreSelect, films}) => {
+const GenreList = ({genres, selectedGenreTab, onGenreSelect, films}) => {
   return (
     <ul className="catalog__genres-list">
       {genres.map((genreItem, i) =>
@@ -11,7 +13,7 @@ const GenreList = ({genres, selectedGenreTab, handleGenreSelect, films}) => {
           <span
             className="catalog__genres-link"
             style={{cursor: `pointer`}}
-            onClick={({target}) => handleGenreSelect(films, target.textContent)}
+            onClick={({target}) => onGenreSelect(films, target.textContent)}
           >{genreItem}
           </span>
         </li>)}
@@ -19,6 +21,19 @@ const GenreList = ({genres, selectedGenreTab, handleGenreSelect, films}) => {
   );
 };
 
+const mapStateToProps = (state) => ({
+  selectedGenreTab: state.selectedGenreTab,
+  films: state.films,
+  genres: state.genres,
+});
+
+const mapDispatchToProps = (dispatch) => ({
+  onGenreSelect(films, genre) {
+    dispatch(ActionCreator.genreSelect(genre));
+    dispatch(ActionCreator.filterFilmsByGenre(films, genre));
+  },
+});
+
 GenreList.propTypes = PropTypes.string.isRequired;
 
-export default GenreList;
+export default connect(mapStateToProps, mapDispatchToProps)(GenreList);
